requirements.txt
google-generativeai
python-dotenv
requests
pandas
